linux build now working; TODO: replace harcoded values in unsuck_platform_specific.cpp

+++ Converter/include/sampler_random.h
@@ -73,9 +73,9 @@ struct SamplerRandom : public Sampler {
 				int64_t y = double(gridSize) * ny;
 				int64_t z = double(gridSize) * nz;
 
-				x = std::max(0ll, std::min(x, gridSize - 1));
-				y = std::max(0ll, std::min(y, gridSize - 1));
-				z = std::max(0ll, std::min(z, gridSize - 1));
+				x = std::max(int64_t(0), std::min(x, gridSize - 1));
+				y = std::max(int64_t(0), std::min(y, gridSize - 1));
+				z = std::max(int64_t(0), std::min(z, gridSize - 1));
 
 				int64_t index = x + y * gridSize + z * gridSize * gridSize;
