+++ Converter/src/indexer_random.cpp
@@ -259,9 +259,9 @@ void Indexer::doSampling(shared_ptr<Node> node, int64_t dbgShouldWriteBytes) {
 			int64_t y = double(gridSize) * ny;
 			int64_t z = double(gridSize) * nz;
 
-			x = std::max(0ll, std::min(x, gridSize - 1));
-			y = std::max(0ll, std::min(y, gridSize - 1));
-			z = std::max(0ll, std::min(z, gridSize - 1));
+			x = std::max(int64_t(0), std::min(x, gridSize - 1));
+			y = std::max(int64_t(0), std::min(y, gridSize - 1));
+			z = std::max(int64_t(0), std::min(z, gridSize - 1));
 
 			int64_t index = x + y * gridSize + z * gridSize * gridSize;
 
@@ -627,9 +627,9 @@ void buildHierarchy(Indexer* indexer, Node* node, shared_ptr<Buffer> points, int
 		int64_t iy = double(counterGridSize) * (y - min.y) / size.y;
 		int64_t iz = double(counterGridSize) * (z - min.z) / size.z;
 
-		ix = std::max(0ll, std::min(ix, counterGridSize - 1));
-		iy = std::max(0ll, std::min(iy, counterGridSize - 1));
-		iz = std::max(0ll, std::min(iz, counterGridSize - 1));
+		ix = std::max(int64_t(0), std::min(ix, counterGridSize - 1));
+		iy = std::max(int64_t(0), std::min(iy, counterGridSize - 1));
+		iz = std::max(int64_t(0), std::min(iz, counterGridSize - 1));
 
 		int64_t index = mortonEncode_magicbits(iz, iy, ix);
